fix: Guild Member Lists & Guild Search lists are now clickable
Previously, the guild member list menus had the rightmost and
bottommost column/row (respectively) not be clickable. This occured
because the contents (usually a Grid) was offset and the Box containing
them was not.

This has been resolved through and admitedly hacky solution where:

1) A box is rendered as before and used to fetch the size of the contents (Grid)
2) A box is rendered whose components are blank cells to clear out step (1)
3) The final box is rendered using fillMaxSize() to ensure that the contents
fit inside it.

This separates the size extraction and the content containing steps.

This ensures that:

- The 'itemsPerPage' or 'itemsPerLine' and 'totalLines' variables
are of the size of the original contents instead of their enclosing Box.

- The enclosing box for the contents is of sufficient size such that all
the elements are clickable.

Alternative solutions could include altering the spec for Scrollable and
Paginated to take in an optional Size override.